+++ src/components/basic/stake/BalanceYFD.tsx
@@ -5,9 +5,7 @@ import convertFromBase from 'utilities/converters/convertFromBase';
 function BalanceYFD() {
   const { tokenBalance } = useContractYFD();
 
-  return (
-    <Text>Balance YFD: {convertFromBase(Number(tokenBalance)).toFixed(5)}</Text>
-  );
+  return <Text>{convertFromBase(Number(tokenBalance)).toFixed(5)}</Text>;
 }
 
 export default BalanceYFD;
