+++ src/components/basic/stake/BalancefYFD.tsx
@@ -5,11 +5,7 @@ import convertFromBase from 'utilities/converters/convertFromBase';
 function BalancefYFD() {
   const { tokenBalance } = useContractForge();
 
-  return (
-    <Text>
-      Balance fYFD: {convertFromBase(Number(tokenBalance)).toFixed(5)}
-    </Text>
-  );
+  return <Text>{convertFromBase(Number(tokenBalance)).toFixed(5)}</Text>;
 }
 
 export default BalancefYFD;
