+++ src/components/pageBody/PageBody.tsx
@@ -1,6 +1,5 @@
 import { Box, Flex, Heading } from '@chakra-ui/react';
 import ProposalsList from './proposal/ProposalsList';
-import StakedYFD from './stake/stakedYFD/StakedYFD';
 
 function PageBody() {
   return (
@@ -9,7 +8,6 @@ function PageBody() {
         <Heading as="h1" size="xl">
           Foundry
         </Heading>
-        <StakedYFD />
         <ProposalsList />
       </Flex>
     </Box>
